Fix bug again

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -10,7 +10,7 @@
 
 int print_last_digit(int n)
 {
-long long int last_dig = abs(n);
+long int last_dig = abs(n);
 
 last_dig %= 10;
 _putchar('0' + last_dig);
